requirements.txt
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
cohere>=5.5.6
openai>=1.50.0
langsmith>=0.1.38
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
